PIX: Preclude debug output of ray query handle from phi or select (since the value isn't really an int, it's a handle) (#6309)
AllocateRayQuery returns an int, but it's not really an int: it's a
handle to the query. Thus, it's not sensible for the PIX debugging
instrumentation to attempt to write it out to the debug-data-UAV. The
previous code did an explicit check against the type of the value to be
written, but if that value is actually a phi itself, then that check
would fail. So now we recursively run through the phi values looking to
see if any of its antecedents are a phi, and if so, refuse to send its
value to the UAV.

(cherry picked from commit 5a3178590581b4a7deea17dd45b6c4f1b51edb53)